+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/SharedInformer.java
@@ -59,4 +59,9 @@ public interface SharedInformer<T> {
    * @return string value
    */
   String lastSyncResourceVersion();
+  
+  /**
+   * Return true if the informer is running
+   */
+  boolean isRunning();
 }