+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/impl/DefaultSharedIndexInformer.java
@@ -230,4 +230,9 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
     }
     return desired < check ? check : desired;
   }
+
+  @Override
+  public boolean isRunning() {
+    return !stopped && started && controller.isRunning(); 
+  }
 }